refactor(#11): 필요없는 import, style-component 삭제, api 비동기 처리
